Added generic reader script to minimise duplication
